requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
pymupdf4llm
python-dotenv
pydantic
langgraph
pytest
